+++ CHANGELOG.md
@@ -29,6 +29,7 @@
   - Add JWT token valiations
   - Fix User authorization profiles
   - add TokenAuthenticationFilterService to security configuration context for HTTP security
+  - add authentication manager builder with bcrypt password encoder
 
 ## 2021.11.17